refactor: 옵션 prop 안전 호출 적용 및 파일명 수정
- 옵션 prop 호출 시 ?.()로 런타임 오류 방지
- 파일명 오타 수정 (KoreanStandaradTime → KoreanStandardTime)
